refactored helperPopup into its own file
